Moved vtk related code to NextSubVolume
